Isolate agent tool-result continuation helpers
The agent runtime index carried pure tool-result collection, continuation, and streamed-tool materialization logic inline with the execution loops. Moving that logic into a focused module keeps the runtime loop thinner while preserving the existing runtime index exports for callers.

Constraint: Public imports from the runtime index must remain stable for this decomposition slice.
Rejected: Split durable conversation state in the same commit | broader surface needs separate regression coverage.
Confidence: high
Scope-risk: narrow
Directive: Keep future runtime decomposition slices behavior-preserving and covered before moving loop-owned state.
Tested: deno test --no-check --allow-all src/agent/runtime/tool-result-continuation.test.ts
Tested: deno test --no-check --allow-all src/agent/runtime
Tested: deno fmt --check src/agent/runtime/index.ts src/agent/runtime/tool-result-continuation.ts src/agent/runtime/tool-result-continuation.test.ts
Tested: deno lint src/agent/runtime/index.ts src/agent/runtime/tool-result-continuation.ts src/agent/runtime/tool-result-continuation.test.ts
Tested: deno check --frozen src/agent/runtime/index.ts src/agent/runtime/tool-result-continuation.ts src/agent/runtime/tool-result-continuation.test.ts
Tested: deno task verify:quick
Not-tested: Hosted CI binary e2e before PR creation
